Conditionally execute tests based on built modules
