+++ cMake/FreeCAD_Helpers/CheckInterModuleDependencies.cmake
@@ -31,7 +31,7 @@ macro(CheckInterModuleDependencies)
     REQUIRES_MODS(BUILD_OPENSCAD           BUILD_MESH_PART BUILD_DRAFT)
     REQUIRES_MODS(BUILD_MATERIAL_EXTERNAL  BUILD_MATERIAL)
     REQUIRES_MODS(BUILD_MEASURE            BUILD_PART)
-    REQUIRES_MODS(BUILD_PART               BUILD_MATERIAL BUILD_SKETCHER BUILD_SPREADSHEET)
+    REQUIRES_MODS(BUILD_PART               BUILD_MATERIAL)
     REQUIRES_MODS(BUILD_PART_DESIGN        BUILD_SKETCHER)
     # REQUIRES_MODS(BUILD_CAM              BUILD_PART BUILD_MESH BUILD_ROBOT)
     REQUIRES_MODS(BUILD_CAM                BUILD_PART BUILD_MESH)
