+++ src/Mod/Sketcher/SketcherTests/TestSketcherSolver.py
@@ -481,218 +481,222 @@ class TestSketcherSolver(unittest.TestCase):
         )
 
     def testRemovedExternalGeometryReference(self):
-        body = self.Doc.addObject("PartDesign::Body", "Body")
-        sketch = body.newObject("Sketcher::SketchObject", "Sketch")
-        CreateRectangleSketch(sketch, (0, 0), (30, 30))
-        pad = body.newObject("PartDesign::Pad", "Pad")
-        pad.Profile = sketch
-        sketch1 = body.newObject("Sketcher::SketchObject", "Sketch1")
-        CreateCircleSketch(sketch1, (15, 15), 0.25)
-        self.Doc.recompute()
-        self.assertEqual(len(pad.Shape.Edges), 12)
+        if "BUILD_PARTDESIGN" in FreeCAD.__cmake__:
+            body = self.Doc.addObject("PartDesign::Body", "Body")
+            sketch = body.newObject("Sketcher::SketchObject", "Sketch")
+            CreateRectangleSketch(sketch, (0, 0), (30, 30))
+            pad = body.newObject("PartDesign::Pad", "Pad")
+            pad.Profile = sketch
+            sketch1 = body.newObject("Sketcher::SketchObject", "Sketch1")
+            CreateCircleSketch(sketch1, (15, 15), 0.25)
+            self.Doc.recompute()
+            self.assertEqual(len(pad.Shape.Edges), 12)
 
-        hole = self.Doc.addObject("PartDesign::Hole", "Hole")
-        hole.Refine = True
-        hole.Reversed = True
-        body.addObject(hole)
-        hole.Profile = sketch1
-        hole.DrillPointAngle = 118.000000
-        hole.Diameter = 6.000000
-        hole.TaperedAngle = 90.000000
-        hole.Tapered = 0
-        hole.Depth = 8.000000
-        hole.Threaded = 1
-        hole.ModelThread = 0
-        hole.ThreadDepthType = 0
-        hole.ThreadType = 1
-        hole.ThreadSize = 16
-        hole.ThreadClass = 0
-        hole.ThreadDirection = 0
-        hole.HoleCutType = 0
-        hole.DepthType = 0
-        hole.DrillPoint = 1
-        hole.DrillForDepth = 0
-        self.Doc.recompute()
-        # 15 edges if it's passthrough-flat 17 if DrillPoint = 1
-        self.assertEqual(len(hole.Shape.Edges), 17)
+            hole = self.Doc.addObject("PartDesign::Hole", "Hole")
+            hole.Refine = True
+            hole.Reversed = True
+            body.addObject(hole)
+            hole.Profile = sketch1
+            hole.DrillPointAngle = 118.000000
+            hole.Diameter = 6.000000
+            hole.TaperedAngle = 90.000000
+            hole.Tapered = 0
+            hole.Depth = 8.000000
+            hole.Threaded = 1
+            hole.ModelThread = 0
+            hole.ThreadDepthType = 0
+            hole.ThreadType = 1
+            hole.ThreadSize = 16
+            hole.ThreadClass = 0
+            hole.ThreadDirection = 0
+            hole.HoleCutType = 0
+            hole.DepthType = 0
+            hole.DrillPoint = 1
+            hole.DrillForDepth = 0
+            self.Doc.recompute()
+            # 15 edges if it's passthrough-flat 17 if DrillPoint = 1
+            self.assertEqual(len(hole.Shape.Edges), 17)
 
-        hole.ModelThread = 1
-        sketch2 = body.newObject("Sketcher::SketchObject", "Sketch2")
-        CreateRectangleSketch(sketch2, (0, 0), (3, 3))
-        self.Doc.recompute()
-        self.assertGreater(len(hole.Shape.Edges), 17)
-        # 77 edges for basic profile
-        self.assertEqual(len(hole.Shape.Edges), 77)
+            hole.ModelThread = 1
+            sketch2 = body.newObject("Sketcher::SketchObject", "Sketch2")
+            CreateRectangleSketch(sketch2, (0, 0), (3, 3))
+            self.Doc.recompute()
+            self.assertGreater(len(hole.Shape.Edges), 17)
+            # 77 edges for basic profile
+            self.assertEqual(len(hole.Shape.Edges), 77)
 
-        # Edges in the thread should disappear when we stop modeling thread
-        sketch2.addExternal("Hole", "Edge29")
-        hole.ModelThread = 0
-        hole.Refine = 1
-        self.Doc.recompute()
-        self.assertEqual(len(hole.Shape.Edges), 17)
-        self.assertEqual(len(sketch2.ExternalGeometry), 0)
+            # Edges in the thread should disappear when we stop modeling thread
+            sketch2.addExternal("Hole", "Edge29")
+            hole.ModelThread = 0
+            hole.Refine = 1
+            self.Doc.recompute()
+            self.assertEqual(len(hole.Shape.Edges), 17)
+            self.assertEqual(len(sketch2.ExternalGeometry), 0)
 
     def testSaveLoadWithExternalGeometryReference(self):
-        # Arrange
-        body = self.Doc.addObject("PartDesign::Body", "Body")
-        sketch = self.Doc.addObject("Sketcher::SketchObject", "Sketch")
-        CreateRectangleSketch(sketch, (0, 0), (30, 30))
-        pad = self.Doc.addObject("PartDesign::Pad", "Pad")
-        pad.Profile = sketch
-        sketch1 = self.Doc.addObject("Sketcher::SketchObject", "Sketch1")
-        body.addObject(sketch)
-        body.addObject(pad)
-        body.addObject(sketch1)
-        self.Doc.recompute()
-        sketch1.addExternal("Pad", "Edge12")
-        self.Doc.recompute()
+        if "BUILD_PARTDESIGN" in FreeCAD.__cmake__:
+            # Arrange
+            body = self.Doc.addObject("PartDesign::Body", "Body")
+            sketch = self.Doc.addObject("Sketcher::SketchObject", "Sketch")
+            CreateRectangleSketch(sketch, (0, 0), (30, 30))
+            pad = self.Doc.addObject("PartDesign::Pad", "Pad")
+            pad.Profile = sketch
+            sketch1 = self.Doc.addObject("Sketcher::SketchObject", "Sketch1")
+            body.addObject(sketch)
+            body.addObject(pad)
+            body.addObject(sketch1)
+            self.Doc.recompute()
+            sketch1.addExternal("Pad", "Edge12")
+            self.Doc.recompute()
 
-        # Act: Try changing sketch before the save
-        sketch = self.Doc.getObject("Sketch")
-        g1 = sketch.Constraints[11].First
-        d1 = sketch.Constraints[11].Value
-        sketch.delConstraint(11)
-        sketch.addConstraint(Sketcher.Constraint("Distance", g1, d1 - 1.0))
-        self.Doc.recompute()
+            # Act: Try changing sketch before the save
+            sketch = self.Doc.getObject("Sketch")
+            g1 = sketch.Constraints[11].First
+            d1 = sketch.Constraints[11].Value
+            sketch.delConstraint(11)
+            sketch.addConstraint(Sketcher.Constraint("Distance", g1, d1 - 1.0))
+            self.Doc.recompute()
 
-        # Act: Save and reload the file
-        filename = self.Doc.Name + ".FCStd"
-        self.Doc.saveAs(filename)
-        FreeCAD.closeDocument(self.Doc.Name)
-        self.Doc = FreeCAD.openDocument(filename)
-        pad = self.Doc.getObject("Pad")
-        sketch1 = self.Doc.getObject("Sketch1")
-        self.Doc.recompute()
+            # Act: Save and reload the file
+            filename = self.Doc.Name + ".FCStd"
+            self.Doc.saveAs(filename)
+            FreeCAD.closeDocument(self.Doc.Name)
+            self.Doc = FreeCAD.openDocument(filename)
+            pad = self.Doc.getObject("Pad")
+            sketch1 = self.Doc.getObject("Sketch1")
+            self.Doc.recompute()
 
-        # Act:  change sketch after restore ( trigger missing references if there is a bug )
-        sketch = self.Doc.getObject("Sketch")
-        g1 = sketch.Constraints[11].First
-        d1 = sketch.Constraints[11].Value
-        sketch.delConstraint(11)
-        sketch.addConstraint(Sketcher.Constraint("Distance", g1, d1 - 1.0))
-        self.Doc.recompute()
+            # Act:  change sketch after restore ( trigger missing references if there is a bug )
+            sketch = self.Doc.getObject("Sketch")
+            g1 = sketch.Constraints[11].First
+            d1 = sketch.Constraints[11].Value
+            sketch.delConstraint(11)
+            sketch.addConstraint(Sketcher.Constraint("Distance", g1, d1 - 1.0))
+            self.Doc.recompute()
 
-        # Assert
-        self.assertEqual(pad.Shape.ElementMapSize, 30)
-        self.assertIn("Edge12", pad.Shape.ElementReverseMap)
-        self.assertIn((pad, ("Edge12",)), sketch1.ExternalGeometry)  # Not "?Edge12"
+            # Assert
+            self.assertEqual(pad.Shape.ElementMapSize, 30)
+            self.assertIn("Edge12", pad.Shape.ElementReverseMap)
+            self.assertIn((pad, ("Edge12",)), sketch1.ExternalGeometry)  # Not "?Edge12"
 
     def testTNPExternalGeometryStored(self):
         # Arrange
-        import xml.etree.ElementTree as ET
+        if "BUILD_PARTDESIGN" in FreeCAD.__cmake__:
+            import xml.etree.ElementTree as ET
 
-        sketch = self.Doc.addObject("Sketcher::SketchObject", "Sketch")
-        CreateRectangleSketch(sketch, (0, 0), (30, 30))
-        sketch1 = self.Doc.addObject("Sketcher::SketchObject", "Sketch1")
-        pad = self.Doc.addObject("PartDesign::Pad", "Pad")
-        pad.Profile = sketch
-        self.Doc.recompute()
-        sketch1.addExternal("Pad", "Edge12")
-        self.Doc.recompute()
-        # Act
-        root = ET.fromstring("<all>" + sketch1.Content + "</all>")
-        # Can't use != in an xpath because it wasn't added until python 3.10.
-        # "*/*[@name='ExternalGeo']//*/[@type='Sketcher::ExternalGeometryExtension']/[@Ref!='']"
-        extRefs = root.findall(
-            "*/*[@name='ExternalGeo']//*/[@type='Sketcher::ExternalGeometryExtension']/[@Ref='']"
-        )
-        extRefsAll = root.findall(
-            "*/*[@name='ExternalGeo']//*/[@type='Sketcher::ExternalGeometryExtension']/[@Ref]"
-        )
-        # Assert
-        self.assertEqual(len(extRefs), 2)
-        self.assertEqual(len(extRefsAll), 3)
-        self.assertEqual(root.tag, "all")
-        # Act
-        filename = self.Doc.Name + ".FCStd"
-        self.Doc.saveAs(filename)
-        FreeCAD.closeDocument(self.Doc.Name)
-        self.Doc = FreeCAD.openDocument(filename)
-        # Assert
-        root = ET.fromstring("<all>" + self.Doc.getObject("Sketch1").Content + "</all>")
-        extRefs = root.findall(
-            "*/*[@name='ExternalGeo']//*/[@type='Sketcher::ExternalGeometryExtension']/[@Ref='']"
-        )
-        extRefsAll = root.findall(
-            "*/*[@name='ExternalGeo']//*/[@type='Sketcher::ExternalGeometryExtension']/[@Ref]"
-        )
-        self.assertEqual(len(extRefs), 2)
-        self.assertEqual(len(extRefsAll), 3)
-        self.assertEqual(root.tag, "all")
-        # Act to change the constraint
-        sketch = self.Doc.getObject("Sketch")
-        g1 = sketch.Constraints[11].First
-        d1 = sketch.Constraints[11].Value
-        sketch.delConstraint(11)
-        sketch.addConstraint(Sketcher.Constraint("Distance", g1, d1 - 1.0))
-        self.Doc.recompute()
-        # Assert
-        root = ET.fromstring("<all>" + self.Doc.getObject("Sketch1").Content + "</all>")
-        extRefs = root.findall(
-            "*/*[@name='ExternalGeo']//*/[@type='Sketcher::ExternalGeometryExtension']/[@Ref='']"
-        )
-        extRefsAll = root.findall(
-            "*/*[@name='ExternalGeo']//*/[@type='Sketcher::ExternalGeometryExtension']/[@Ref]"
-        )
-        self.assertEqual(len(extRefs), 2)
-        self.assertEqual(len(extRefsAll), 3)
-        self.assertEqual(root.tag, "all")
+            sketch = self.Doc.addObject("Sketcher::SketchObject", "Sketch")
+            CreateRectangleSketch(sketch, (0, 0), (30, 30))
+            sketch1 = self.Doc.addObject("Sketcher::SketchObject", "Sketch1")
+            pad = self.Doc.addObject("PartDesign::Pad", "Pad")
+            pad.Profile = sketch
+            self.Doc.recompute()
+            sketch1.addExternal("Pad", "Edge12")
+            self.Doc.recompute()
+            # Act
+            root = ET.fromstring("<all>" + sketch1.Content + "</all>")
+            # Can't use != in an xpath because it wasn't added until python 3.10.
+            # "*/*[@name='ExternalGeo']//*/[@type='Sketcher::ExternalGeometryExtension']/[@Ref!='']"
+            extRefs = root.findall(
+                "*/*[@name='ExternalGeo']//*/[@type='Sketcher::ExternalGeometryExtension']/[@Ref='']"
+            )
+            extRefsAll = root.findall(
+                "*/*[@name='ExternalGeo']//*/[@type='Sketcher::ExternalGeometryExtension']/[@Ref]"
+            )
+            # Assert
+            self.assertEqual(len(extRefs), 2)
+            self.assertEqual(len(extRefsAll), 3)
+            self.assertEqual(root.tag, "all")
+            # Act
+            filename = self.Doc.Name + ".FCStd"
+            self.Doc.saveAs(filename)
+            FreeCAD.closeDocument(self.Doc.Name)
+            self.Doc = FreeCAD.openDocument(filename)
+            # Assert
+            root = ET.fromstring("<all>" + self.Doc.getObject("Sketch1").Content + "</all>")
+            extRefs = root.findall(
+                "*/*[@name='ExternalGeo']//*/[@type='Sketcher::ExternalGeometryExtension']/[@Ref='']"
+            )
+            extRefsAll = root.findall(
+                "*/*[@name='ExternalGeo']//*/[@type='Sketcher::ExternalGeometryExtension']/[@Ref]"
+            )
+            self.assertEqual(len(extRefs), 2)
+            self.assertEqual(len(extRefsAll), 3)
+            self.assertEqual(root.tag, "all")
+            # Act to change the constraint
+            sketch = self.Doc.getObject("Sketch")
+            g1 = sketch.Constraints[11].First
+            d1 = sketch.Constraints[11].Value
+            sketch.delConstraint(11)
+            sketch.addConstraint(Sketcher.Constraint("Distance", g1, d1 - 1.0))
+            self.Doc.recompute()
+            # Assert
+            root = ET.fromstring("<all>" + self.Doc.getObject("Sketch1").Content + "</all>")
+            extRefs = root.findall(
+                "*/*[@name='ExternalGeo']//*/[@type='Sketcher::ExternalGeometryExtension']/[@Ref='']"
+            )
+            extRefsAll = root.findall(
+                "*/*[@name='ExternalGeo']//*/[@type='Sketcher::ExternalGeometryExtension']/[@Ref]"
+            )
+            self.assertEqual(len(extRefs), 2)
+            self.assertEqual(len(extRefsAll), 3)
+            self.assertEqual(root.tag, "all")
 
     def testConstructionToggleTNP(self):
         """Bug 15484"""
-        # Arrange
-        import xml.etree.ElementTree as ET
+        if "BUILD_PARTDESIGN" in FreeCAD.__cmake__:
+            # Arrange
+            import xml.etree.ElementTree as ET
 
-        body = self.Doc.addObject("PartDesign::Body", "Body")
-        sketch = self.Doc.addObject("Sketcher::SketchObject", "Sketch")
-        CreateRectangleSketch(sketch, (0, 0), (30, 30))
-        # Add zigsag geo as construction lines
-        i = sketch.GeometryCount
-        sketch.addGeometry(
-            Part.LineSegment(FreeCAD.Vector(0, 0, 0), FreeCAD.Vector(-10, 10, 0)), True
-        )
-        sketch.addGeometry(
-            Part.LineSegment(FreeCAD.Vector(-10, 10, 0), FreeCAD.Vector(-5, 20, 0)), True
-        )
-        sketch.addGeometry(
-            Part.LineSegment(FreeCAD.Vector(-5, 20, 0), FreeCAD.Vector(-10, 25, 0)), True
-        )
-        sketch.addGeometry(
-            Part.LineSegment(FreeCAD.Vector(-10, 25, 0), FreeCAD.Vector(0, 30, 0)), True
-        )
-        sketch.addConstraint(Sketcher.Constraint("Coincident", i + 0, 2, i + 1, 1))
-        sketch.addConstraint(Sketcher.Constraint("Coincident", i + 1, 2, i + 2, 1))
-        sketch.addConstraint(Sketcher.Constraint("Coincident", i + 2, 2, i + 3, 1))
-        sketch.addConstraint(Sketcher.Constraint("Coincident", i + 3, 2, 0, 1))
-        sketch.addConstraint(Sketcher.Constraint("Coincident", i + 0, 1, 2, 2))
+            body = self.Doc.addObject("PartDesign::Body", "Body")
+            sketch = self.Doc.addObject("Sketcher::SketchObject", "Sketch")
+            CreateRectangleSketch(sketch, (0, 0), (30, 30))
+            # Add zigsag geo as construction lines
+            i = sketch.GeometryCount
+            sketch.addGeometry(
+                Part.LineSegment(FreeCAD.Vector(0, 0, 0), FreeCAD.Vector(-10, 10, 0)), True
+            )
+            sketch.addGeometry(
+                Part.LineSegment(FreeCAD.Vector(-10, 10, 0), FreeCAD.Vector(-5, 20, 0)), True
+            )
+            sketch.addGeometry(
+                Part.LineSegment(FreeCAD.Vector(-5, 20, 0), FreeCAD.Vector(-10, 25, 0)), True
+            )
+            sketch.addGeometry(
+                Part.LineSegment(FreeCAD.Vector(-10, 25, 0), FreeCAD.Vector(0, 30, 0)), True
+            )
+            sketch.addConstraint(Sketcher.Constraint("Coincident", i + 0, 2, i + 1, 1))
+            sketch.addConstraint(Sketcher.Constraint("Coincident", i + 1, 2, i + 2, 1))
+            sketch.addConstraint(Sketcher.Constraint("Coincident", i + 2, 2, i + 3, 1))
+            sketch.addConstraint(Sketcher.Constraint("Coincident", i + 3, 2, 0, 1))
+            sketch.addConstraint(Sketcher.Constraint("Coincident", i + 0, 1, 2, 2))
 
-        pad = self.Doc.addObject("PartDesign::Pad", "Pad")
-        pad.Profile = sketch
-        body.addObject(sketch)
-        body.addObject(pad)
-        self.Doc.recompute()
-        sketch1 = self.Doc.addObject("Sketcher::SketchObject", "Sketch1")
-        sketch1.AttachmentSupport = (pad, ("Face6"))
-        sketch1.MapMode = "FlatFace"
-        self.Doc.recompute()
+            pad = self.Doc.addObject("PartDesign::Pad", "Pad")
+            pad.Profile = sketch
+            body.addObject(sketch)
+            body.addObject(pad)
+            self.Doc.recompute()
+            sketch1 = self.Doc.addObject("Sketcher::SketchObject", "Sketch1")
+            sketch1.AttachmentSupport = (pad, ("Face6"))
+            sketch1.MapMode = "FlatFace"
+            self.Doc.recompute()
 
-        CreateCircleSketch(sketch1, (2, 2, 0), 1)
-        CreateCircleSketch(sketch1, (6, 2, 0), 1)
-        body.addObject(sketch1)
-        self.Doc.recompute()
-        # Act toggle construction lines on in sketch; pad now has 9 instead of 6 faces.
-        sketch.setConstruction(4, False)
-        sketch.setConstruction(5, False)
-        sketch.setConstruction(6, False)
-        sketch.setConstruction(7, False)
-        sketch.setConstruction(3, True)
-        self.Doc.recompute()
-        # Assert
-        # AttachmentSupport is a list of (object,(subobject list)) with 1 entry.  Get the
-        # first and only subobject name in second part of that first tuple and see that it moved
-        # from the Face6 we set above.
-        self.assertEqual(sketch1.AttachmentSupport[0][1][0], "Face9")
-        self.assertIn("Face6", pad.Shape.ElementReverseMap)  # different Face6 exists
+            CreateCircleSketch(sketch1, (2, 2, 0), 1)
+            CreateCircleSketch(sketch1, (6, 2, 0), 1)
+            body.addObject(sketch1)
+            self.Doc.recompute()
+            # Act toggle construction lines on in sketch; pad now has 9 instead of 6 faces.
+            sketch.setConstruction(4, False)
+            sketch.setConstruction(5, False)
+            sketch.setConstruction(6, False)
+            sketch.setConstruction(7, False)
+            sketch.setConstruction(3, True)
+            self.Doc.recompute()
+            # Assert
+            # AttachmentSupport is a list of (object,(subobject list)) with 1 entry.  Get the
+            # first and only subobject name in second part of that first tuple and see that it moved
+            # from the Face6 we set above.
+            self.assertEqual(sketch1.AttachmentSupport[0][1][0], "Face9")
+            self.assertIn("Face6", pad.Shape.ElementReverseMap)  # different Face6 exists
 
     # TODO other tests:
     #  getHigherElement
